+++ src/tests/cairo_run_test.rs
@@ -795,6 +795,13 @@ fn is_zero() {
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn cairo_run_secp256r1_div_mod_n() {
+    let program_data = include_bytes!("../../cairo_programs/secp256r1_div_mod_n.json");
+    run_program_simple(program_data.as_slice());
+}
+
+#[test]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn is_zero_pack() {
     let program_data = include_bytes!("../../cairo_programs/is_zero_pack.json");
     run_program_simple(program_data.as_slice());